+++ GameServer/gameobjects/Atlantis/Teleporter/AncientBoundDjinn.cs
@@ -50,7 +50,7 @@ namespace DOL.GS
             CurrentRegion = djinnStone.CurrentRegion;
             Heading = djinnStone.Heading;
             Realm = eRealm.None;
-            Flags ^= (uint)GameNPC.eFlags.FLYING | (uint)GameNPC.eFlags.PEACE;
+            Flags ^= GameNPC.eFlags.FLYING | GameNPC.eFlags.PEACE;
             X = djinnStone.X;
             Y = djinnStone.Y;
             Z = djinnStone.Z + HoverHeight;
